+++ zebra/zserv.h
@@ -149,7 +149,6 @@ extern unsigned int multipath_num;
 
 /* Prototypes. */
 extern void zserv_init(void);
-extern void zebra_init(void);
 extern void zebra_if_init(void);
 extern void zebra_zserv_socket_init(char *path);
 extern void hostinfo_get(void);
